Extracted and Refactored Post Header and Post Body
